# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "sakalWeb Information System v5 RESTful API documentation",
        "title": "swis-api v5",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://savla.dev/swapi",
            "email": "info@savla.dev"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/savla-dev/swis-api/blob/master/LICENSE"
        },
        "version": "5.2.4"
    },
    "host": "swapi.savla.su:8049",
    "basePath": "/",
    "paths": {
...
        "/users": {
            "get": {
                "description": "get users complete list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/users.Users"
                        }
                    }
                }
            },
            "post": {
                "description": "add new user to users array",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Add new user to users array",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/users.User"
                        }
                    }
                }
            }
        },
...
        "/users/{name}/active": {
            "put": {
                "description": "toggle active boolean for {user}",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Toggle active boolean for {user}",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/users.User"
                        }
                    }
                }
            }
        },
        "/users/{name}/keys/ssh": {
            "get": {
                "description": "fetch :user ssh key array output in plain text",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get User's SSH keys in plain text",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/users.User"
                        }
                    }
                }
            },
            "post": {
                "description": "add new SSH keys to :user param",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Add SSH public keys to User",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/users.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "users.User": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "active": {
                    "description": "Presence/Absence boolean. If false, one is not allowed to log-in (token is rejected),\nto interract with savla-dev infra in general (by default).",
                    "type": "boolean"
                },
                "country": {
                    "description": "Country of origin -- to help maintain global contacts.",
                    "type": "string"
                },
                "discord_username": {
                    "description": "Discord account/profile name.",
                    "type": "string"
                },
                "email_alias": {
                    "description": "Email alias in cloudflare email routing",
                    "type": "string"
                },
                "email_main": {
                    "description": "Email address main, personal",
                    "type": "string"
                },
                "full_name": {
                    "description": "Full Name of such user.",
                    "type": "string"
                },
                "gdpr_consent": {
                    "description": "Important GDPR consent boolean -- if false, user's details should be omitted!\nSEE more -- https://gdpr.eu/checklist/",
                    "type": "boolean"
                },
                "github_username": {
                    "description": "GitHub account/profile name (used for SSH public keys importing).",
                    "type": "string"
                },
                "gpg_keys": {
                    "description": "User's GPG public keys array.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "ID not used anymore as indexing is used differently now (searching by Name, index respects array implicit property).\nID       \tstring \t\t`json:\"id\"`",
                    "type": "string"
                },
                "roles": {
                    "description": "User's given roles -- a role labels array.\nRoles []roles.Role `json:\"roles\"`",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "spotify_link": {
                    "description": "Spotify link to one's profile.",
                    "type": "string"
                },
                "ssh_keys": {
                    "description": "User's SSH public keys array.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "token_hmac": {
                    "description": "Unique token used for auth purposes, SHA?-HMAC.",
                    "type": "string"
                },
                "wireguard_vpn": {
                    "description": "All Wireguard config objects -- an array.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/users.Wireguard"
                    }
                }
            }
        },
        "users.Users": {
            "type": "object",
            "properties": {
                "users": {
                    "description": "An array of User objects",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/users.User"
                    }
                }
            }
        },
        "users.Wireguard": {
            "type": "object",
            "properties": {
                "allowed_ips": {
                    "description": "Allowed IP address(es) list on the side of server (vpn_gateway_server).",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "device_name": {
                    "description": "Unique device name (for such user).",
                    "type": "string"
                },
                "ip_address": {
                    "description": "User's private IP address.",
                    "type": "string"
                },
                "permission": {
                    "description": "Is the user given permission to dial a connection?",
                    "type": "boolean"
                },
                "private_key": {
                    "description": "Wireguard private key.\nTODO: should be encrypted?",
                    "type": "string"
                },
                "public_key": {
                    "description": "Wireguard public key.",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "apiKey": {
            "type": "apiKey",
            "name": "X-Auth-Token",
            "in": "header"
        }
    }
}